requirements.txt
Flask==2.1.0
mysql-connector-python==8.0.25
